refactor: change quantity type from int to double

+++ src/main/java/com/brotes/api/modelo/itemPedido/DatosItemPedido.java
@@ -1,4 +1,4 @@
 package com.brotes.api.modelo.itemPedido;
 
-public record DatosItemPedido(int cantidad, Long idPedido, Long idProducto) {
+public record DatosItemPedido(Double cantidad, Long idPedido, Long idProducto) {
 }
